refactor: moved rest of variables from constants.ts
