+++ sites/geohub/src/components/RasterLayer.svelte
@@ -4,7 +4,7 @@
   import RasterTransform from '$components/controls/RasterTransform.svelte'
   import LayerNameGroup from '$components/control-groups/LayerNameGroup.svelte'
   import OpacityPanel from '$components/controls/OpacityPanel.svelte'
-  import { TabNames } from '$lib/constants'
+  import { TabNames } from '$lib/config/AppConfig'
   import type { Layer, RasterTileMetadata } from '$lib/types'
   import RasterHistogram from '$components/controls/RasterHistogram.svelte'
   import { Tabs } from '@undp-data/svelte-undp-design'
