+++ sites/geohub/src/components/__tests__/ColorMapPickerCard.test.ts
@@ -2,7 +2,7 @@ import { describe, beforeEach, expect, it } from 'vitest'
 import { cleanup, render, within, type RenderResult } from '@testing-library/svelte'
 
 import ColorMapPickerCard from '$components/controls/ColorMapPickerCard.svelte'
-import { ColorMapTypes } from '$lib/constants'
+import { ColorMapTypes } from '$lib/config/AppConfig'
 import { getRandomColormap } from '$lib/helper'
 
 beforeEach(cleanup)
