+++ sites/geohub/src/lib/config/AppConfig/ClassificationMethod.ts
@@ -1,0 +1,13 @@
+export enum ClassificationMethodTypes {
+  EQUIDISTANT = 'e',
+  QUANTILE = 'q',
+  LOGARITHMIC = 'l',
+  NATURAL_BREAK = 'n',
+}
+
+export enum ClassificationMethodNames {
+  EQUIDISTANT = 'Equidistant',
+  QUANTILE = 'Quantile',
+  LOGARITHMIC = 'Logarithmic',
+  NATURAL_BREAK = 'Natural Breaks',
+}
