+++ sites/geohub/src/lib/config/AppConfig/ColorMapTypes.ts
@@ -1,0 +1,5 @@
+export enum ColorMapTypes {
+  SEQUENTIAL = 'sequential',
+  DIVERGING = 'diverging',
+  QUALITATIVE = 'qualitative',
+}
