+++ sites/geohub/src/lib/config/AppConfig/LegendTypes.ts
@@ -1,0 +1,4 @@
+export enum LegendTypes {
+  DEFAULT = 'default',
+  CLASSIFY = 'classify',
+}
