+++ sites/geohub/src/lib/config/AppConfig/RasterArithmeticOperators.ts
@@ -1,0 +1,8 @@
+export const RasterArithmeticOperators = [
+  { value: '+', label: 'Plus', text: 'addition', disabled: false, symbol: '+' },
+  { value: '-', label: 'Minus', text: 'subtraction', disabled: false, symbol: '-' },
+  { value: '*', label: 'Multiply', text: 'multiplication', disabled: false, symbol: '*' },
+  { value: '/', label: 'Divide', text: 'division', disabled: false, symbol: '/' },
+  { value: '%', label: 'Modulo', text: 'remainder of division', disabled: false, symbol: '%' },
+  { value: '**', label: 'Power', text: 'raise to power', disabled: false, symbol: '^' },
+]
