+++ sites/geohub/src/lib/config/AppConfig/RasterComparisonOperators.ts
@@ -1,0 +1,8 @@
+export const RasterComparisonOperators = [
+  { value: '==', label: 'Equals', text: 'equal to', disabled: false, symbol: '=' },
+  { value: '!=', label: 'Differs', text: 'different then', disabled: false, symbol: '≠' },
+  { value: '>', label: 'Larger', text: 'larger than', disabled: false, symbol: '>' },
+  { value: '<', label: 'Smaller', text: 'smaller than', disabled: false, symbol: '<' },
+  { value: '>=', label: 'Larger or equal', text: 'larger than or equal to', disabled: false, symbol: '>=' },
+  { value: '<=', label: 'Smaller or equal', text: 'smaller than or equal to', disabled: false, symbol: '<=' },
+]
