+++ sites/geohub/src/lib/config/AppConfig/TabNames.ts
@@ -1,0 +1,14 @@
+export enum TabNames {
+  ANALYZE = 'Analyze',
+  DATA = 'Data',
+  LABEL = 'Label',
+  LAYERS = 'Layers',
+  LEGEND = 'Legend',
+  OPACITY = 'Opacity',
+  TRANSFORM = 'Transform',
+  STYLEJSON = 'Specification',
+  ZOOM = 'Zoom',
+  HISTOGRAM = 'Histogram',
+  FILTER = 'Filter',
+  SIMULATION = 'Simulation',
+}
