+++ sites/geohub/src/lib/config/AppConfig/VectorApplyToTypes.ts
@@ -1,0 +1,4 @@
+export enum VectorApplyToTypes {
+  COLOR = 'color',
+  SIZE = 'size',
+}
