+++ sites/geohub/src/lib/config/AppConfig/index.ts
@@ -1,12 +1,15 @@
 export * from './AcceptedExtensions'
 export * from './AccessLevel'
 export * from './attribution'
+export * from './ClassificationMethod'
+export * from './ColorMapTypes'
 export * from './DataCategories'
 export * from './DatasetSearchQueryParams'
 export * from './DatasetSortingColumns'
 export * from './FetchTimeoutMsec'
 export * from './FooterItems'
 export * from './HeaderItems'
+export * from './LegendTypes'
 export * from './LimitOptions'
 export * from './MapSortingColumns'
 export * from './MapStyles'
@@ -14,8 +17,13 @@ export * from './NumberOfClassesMaximum'
 export * from './NumberOfClassesMinimum'
 export * from './NumberOfRandomSamplingPoints'
 export * from './Permission'
+export * from './RasterArithmeticOperators'
+export * from './RasterComparisonOperators'
 export * from './StacMinimumZoom'
+export * from './TabNames'
 export * from './TagInputValues'
 export * from './TagSearchKeys'
 export * from './TokenExpiryPeriodMsec'
 export * from './UniqueValueThreshold'
+export * from './VectorApplyToTypes'
+export * from './VectorFilterOperators'
